# openapi-export.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InterviewCoach",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://protozoan-ripcord-flying.ngrok-free.dev"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/interview/questions": {
      "post": {
        "summary": "Interview Questions",
        "operationId": "interview_questions_interview_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/{role}": {
      "get": {
        "summary": "Get Knowledge",
        "operationId": "get_knowledge_knowledge__role__get",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/skill-gap": {
      "post": {
        "summary": "Skill Gap",
        "operationId": "skill_gap_skill_gap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillGapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/study-plan": {
      "post": {
        "summary": "Study Plan",
        "operationId": "study_plan_study_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/interview-prep": {
      "post": {
        "summary": "Interview Prep",
        "operationId": "interview_prep_interview_prep_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewPrepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/study-plan-from-skills": {
      "post": {
        "summary": "Study Plan From Skills",
        "operationId": "study_plan_from_skills_study_plan_from_skills_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyPlanFromSkillsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api-info": {
      "get": {
        "summary": "Api Info",
        "operationId": "api_info_api_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterviewPrepRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "experience_level": {
            "type": "string",
            "title": "Experience Level"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          }
        },
        "type": "object",
        "required": [
          "role",
          "experience_level",
          "skills"
        ],
        "title": "InterviewPrepRequest"
      },
      "InterviewRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "experience_level": {
            "type": "string",
            "title": "Experience Level"
          }
        },
        "type": "object",
        "required": [
          "role",
          "experience_level"
        ],
        "title": "InterviewRequest"
      },
      "SkillGapRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          }
        },
        "type": "object",
        "required": [
          "role",
          "skills"
        ],
        "title": "SkillGapRequest"
      },
      "StudyPlanFromSkillsRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          }
        },
        "type": "object",
        "required": [
          "role",
          "skills"
        ],
        "title": "StudyPlanFromSkillsRequest"
      },
      "StudyPlanRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills"
          }
        },
        "type": "object",
        "required": [
          "role",
          "missing_skills"
        ],
        "title": "StudyPlanRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}